refactor: rename pubusb
