+++ anvil/server/Cargo.toml
@@ -21,6 +21,7 @@ tracing = "0.1"
 # async
 parking_lot = "0.12"
 futures = "0.3"
+parity-tokio-ipc = { version = "0.9.0", optional = true }
 
 # misc
 serde_json = "1.0.67"
@@ -31,3 +32,7 @@ clap = { version = "3.0.10", features = [
     "derive",
     "env",
 ], optional = true }
+
+[features]
+default = ["ipc"]
+ipc = ["parity-tokio-ipc"]
