+++ anvil/server/src/lib.rs
@@ -20,10 +20,15 @@ use tower_http::{cors::CorsLayer, trace::TraceLayer};
 use tracing::{error, trace};
 
 mod config;
+
+// #[cfg(feature = "ipc")]
 /// handlers for axum server
 mod handler;
+mod ipc;
+mod pubsub;
 mod ws;
-pub use crate::ws::{WsContext, WsRpcHandler};
+
+pub use crate::pubsub::{PubSubContext, PubSubRpcHandler};
 pub use config::ServerConfig;
 
 /// Type alias for the configured axum server
@@ -39,7 +44,7 @@ pub fn serve_http_ws<Http, Ws>(
 ) -> AnvilServer
 where
     Http: RpcHandler,
-    Ws: WsRpcHandler,
+    Ws: PubSubRpcHandler,
 {
     let ServerConfig { allow_origin, no_cors } = config;
